requirements.txt
Flask==3.1.0
flask-cors
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
prometheus-flask-exporter
furl
pillow
grpcio-tools
grpcio
supervisor